+++ scripts/rebuild_specs.js
@@ -4,67 +4,57 @@
 Error.stackTraceLimit = 50;
 import ts from "typescript";
 import fs from "fs";
-import path, { basename } from "path";
+import { rm } from "fs/promises";
+import path, { basename, join } from "path";
 import * as td from "../dist/index.js";
 import { getExpandedEntryPointsForPaths } from "../dist/lib/utils/index.js";
 import { ok } from "assert";
 import { fileURLToPath } from "url";
 import { diagnostics } from "../dist/lib/utils/loggers.js";
+import { getConverter2App, getConverter2Project, getConverterApp } from "../dist/test/programs.js";
+import { TestRouter, TestTheme } from "../dist/test/renderer/testRendererUtils.js";
 
 const base = path.join(
     fileURLToPath(import.meta.url),
     "../../src/test/converter",
 );
 
-async function getApp() {
-    const app = await td.Application.bootstrap(
-        {
-            name: "typedoc",
-            excludeExternals: true,
-            disableSources: false,
-            excludePrivate: false,
-            tsconfig: path.join(base, "tsconfig.json"),
-            externalPattern: ["**/node_modules/**"],
-            entryPointStrategy: td.EntryPointStrategy.Expand,
-            logLevel: td.LogLevel.Warn,
-            gitRevision: "fake",
-            readme: "none",
-        },
-        [new td.TSConfigReader()],
-    );
-    app.serializer.addSerializer({
-        priority: -1,
-        supports(obj) {
-            return obj instanceof td.SourceReference;
-        },
-        /**
-         * @param {td.SourceReference} ref
-         */
-        toObject(ref, obj) {
-            if (obj.url) {
-                obj.url = `typedoc://${
-                    obj.url.substring(
-                        obj.url.indexOf(ref.fileName),
-                    )
-                }`;
-            }
-            return obj;
-        },
-    });
-    app.serializer.addSerializer({
-        priority: -1,
-        supports(obj) {
-            return obj instanceof td.ProjectReflection;
-        },
-        /** @param {td.ProjectReflection} obj */
-        toObject(_refl, obj) {
-            delete obj.packageVersion;
-            return obj;
-        },
-    });
+const app = getConverterApp();
+const app2 = getConverter2App();
 
-    return app;
-}
+app2.renderer.defineTheme("test-theme", TestTheme);
+app2.renderer.defineRouter("test-router", TestRouter);
+
+app.serializer.addSerializer({
+    priority: -1,
+    supports(obj) {
+        return obj instanceof td.SourceReference;
+    },
+    /**
+     * @param {td.SourceReference} ref
+     */
+    toObject(ref, obj) {
+        if (obj.url) {
+            obj.url = `typedoc://${
+                obj.url.substring(
+                    obj.url.indexOf(ref.fileName),
+                )
+            }`;
+        }
+        return obj;
+    },
+});
+app.serializer.addSerializer({
+    priority: -1,
+    supports(obj) {
+        return obj instanceof td.ProjectReflection;
+    },
+    /** @param {td.ProjectReflection} obj */
+    toObject(_refl, obj) {
+        delete obj.packageVersion;
+        return obj;
+    },
+});
 
 /** @type {[string, (app: td.Application) => void, (app: td.Application) => void][]} */
 const conversions = [
@@ -91,10 +81,9 @@ const conversions = [
 
 /**
  * Rebuilds the converter specs for the provided dirs.
- * @param {td.Application} app
  * @param {string[]} dirs
  */
-function rebuildConverterTests(app, dirs) {
+function rebuildConverterTests(dirs) {
     const program = ts.createProgram(app.options.getFileNames(), {
         ...app.options.getCompilerOptions(),
         noEmit: true,
@@ -136,23 +125,38 @@ function rebuildConverterTests(app, dirs) {
     }
 }
 
-async function main(filter = "") {
-    console.log("Base directory is", base);
-    const dirs = await fs.promises.readdir(base, { withFileTypes: true });
-    const app = await getApp();
+/** @param {string} specPath */
+async function rebuildRendererTest(specPath) {
+    console.log(specPath);
+    await rm(specPath, { recursive: true, force: true });
 
-    await rebuildConverterTests(
-        app,
-        dirs
-            .filter((dir) => {
-                if (!dir.isDirectory()) return false;
-                return dir.name.endsWith(filter);
-            })
-            .map((dir) => path.join(base, dir.name)),
-    );
+    app2.options.setValue("theme", "test-theme");
+    app2.options.setValue("router", "test-router");
+    app2.options.setValue("disableSources", true);
+    app2.options.setValue("disableGit", true);
+    app2.options.setValue("sourceLinkTemplate", "{path}:{line}");
+
+    const project = getConverter2Project(["renderer"], ".");
+    project.readme = [{ kind: "text", text: "Readme text" }];
+    await app2.generateDocs(project, specPath);
+    await rm(`${specPath}/assets`, { recursive: true });
+    await rm(`${specPath}/.nojekyll`);
 }
 
-main(process.argv[2]).catch((reason) => {
+async function main() {
+    console.log("Base directory is", base);
+    const dirs = await fs.promises.readdir(base, { withFileTypes: true });
+
+    await rebuildConverterTests(
+        dirs
+            .filter((dir) => dir.isDirectory())
+            .map((dir) => path.join(base, dir.name)),
+    );
+
+    await rebuildRendererTest(join(process.cwd(), "src/test/renderer/specs"));
+}
+
+main().catch((reason) => {
     console.error(reason);
     process.exit(1);
 });