+++ src/lib/converter/types.ts
@@ -35,7 +35,6 @@ import { convertParameterNodes, convertTypeParameterNodes, createSignature } fro
 import { convertSymbol } from "./symbols.js";
 import { isObjectType, isTypeReference } from "./utils/nodes.js";
 import { removeUndefined } from "./utils/reflections.js";
-import { createSymbolReference } from "./factories/types.js";
 import { createSymbolId } from "./factories/symbol-id.js";
 
 export interface TypeConverter<
@@ -340,7 +339,7 @@ const exprWithTypeArgsConverter: TypeConverter<
             );
         }
         const parameters = node.typeArguments?.map((type) => convertType(context, type)) ?? [];
-        const ref = createSymbolReference(
+        const ref = context.createSymbolReference(
             context.resolveAliasedSymbol(targetSymbol),
             context,
         );
@@ -439,7 +438,7 @@ const importType: TypeConverter<ts.ImportTypeNode> = {
             return new IntrinsicType("any");
         }
 
-        return createSymbolReference(
+        return context.createSymbolReference(
             context.resolveAliasedSymbol(symbol),
             context,
             name,
@@ -448,7 +447,7 @@ const importType: TypeConverter<ts.ImportTypeNode> = {
     convertType(context, type) {
         const symbol = type.getSymbol();
         assert(symbol, "Missing symbol when converting import type"); // Should be a compiler error
-        return createSymbolReference(
+        return context.createSymbolReference(
             context.resolveAliasedSymbol(symbol),
             context,
             "__module",
@@ -710,7 +709,7 @@ const queryConverter: TypeConverter<ts.TypeQueryNode> = {
             );
         }
 
-        const ref = createSymbolReference(
+        const ref = context.createSymbolReference(
             context.resolveAliasedSymbol(querySymbol),
             context,
             node.exprName.getText(),
@@ -730,7 +729,7 @@ const queryConverter: TypeConverter<ts.TypeQueryNode> = {
                 )
             }. This is a bug.`,
         );
-        const ref = createSymbolReference(
+        const ref = context.createSymbolReference(
             context.resolveAliasedSymbol(symbol),
             context,
         );
@@ -757,14 +756,13 @@ const referenceConverter: TypeConverter<
         }
 
         const symbol = context.expectSymbolAtLocation(node.typeName);
+        const name = node.typeName.getText();
 
         // Ignore @inline if there are type arguments, as they won't be resolved
         // in the type we just retrieved from node.typeName.
         if (
             !node.typeArguments &&
-            context
-                .getComment(symbol, ReflectionKind.Interface)
-                ?.hasModifier("@inline")
+            context.shouldInline(symbol, name)
         ) {
             // typeLiteralConverter doesn't use the node, so we can get away with lying here.
             // This might not actually be safe, it appears that it is in the relatively small
@@ -778,9 +776,7 @@ const referenceConverter: TypeConverter<
             );
         }
 
-        const name = node.typeName.getText();
-
-        const ref = createSymbolReference(
+        const ref = context.createSymbolReference(
             context.resolveAliasedSymbol(symbol),
             context,
             name,
@@ -804,11 +800,14 @@ const referenceConverter: TypeConverter<
             return ref;
         }
 
-        if (
-            context
-                .getComment(symbol, ReflectionKind.Interface)
-                ?.hasModifier("@inline")
-        ) {
+        let name: string;
+        if (ts.isIdentifier(node.typeName)) {
+            name = node.typeName.text;
+        } else {
+            name = node.typeName.right.text;
+        }
+
+        if (context.shouldInline(symbol, name)) {
             // typeLiteralConverter doesn't use the node, so we can get away with lying here.
             // This might not actually be safe, it appears that it is in the relatively small
             // amount of testing I've done with it, but I wouldn't be surprised if someone manages
@@ -821,14 +820,7 @@ const referenceConverter: TypeConverter<
             );
         }
 
-        let name: string;
-        if (ts.isIdentifier(node.typeName)) {
-            name = node.typeName.text;
-        } else {
-            name = node.typeName.right.text;
-        }
-
-        const ref = createSymbolReference(
+        const ref = context.createSymbolReference(
             context.resolveAliasedSymbol(symbol),
             context,
             name,