+++ src/test/behavior.c2.test.ts
@@ -11,7 +11,7 @@ import {
 import { filterMap } from "#utils";
 import { CommentStyle } from "../lib/utils/options/declaration.js";
 import { TestLogger } from "./TestLogger.js";
-import { getComment, getSigComment, query, querySig } from "./utils.js";
+import { getComment, getSigComment, query, querySig, reflToTree } from "./utils.js";
 import { getConverter2App, getConverter2Project } from "./programs.js";
 
 type NameTree = { [name: string]: NameTree | undefined };
@@ -1203,6 +1203,33 @@ describe("Behavior Tests", () => {
         );
     });
 
+    it("Handles the @preventInline tag preventing @inline from taking effect", () => {
+        const project = convert("inlineTag");
+
+        const bar2 = querySig(project, "bar2");
+        equal(
+            bar2.parameters?.map((p) => p.type?.toString()),
+            ["Record<string, Foo>"],
+        );
+
+        const baz = querySig(project, "Class.baz");
+        equal(
+            baz.parameters?.map((p) => p.type?.toString()),
+            ["Foo"],
+        );
+        equal(baz.type?.toString(), "Complex<number>");
+    });
+
+    it("Handles the @inlineType tag to inline a requested type", () => {
+        const project = convert("inlineTag");
+
+        const selectiveInline = querySig(project, "selectiveInline");
+        equal(
+            selectiveInline.parameters?.map((p) => p.type?.toString()),
+            ["{ inlined: false }"],
+        );
+    });
+
     it("Handles the @useDeclaredType tag on types", () => {
         const project = convert("useDeclaredTypeTag");
         const data = query(project, "Data");
@@ -1401,5 +1428,20 @@ describe("Behavior Tests", () => {
         const project = convert("asConstEnum");
         // Previously, would find TypeDoc's root readme
         equal(project.readme, undefined);
+    });
+
+    it("Supports @document tags", () => {
+        const project = convert("documentTag");
+
+        equal(reflToTree(project), {
+            HasDescriptor: "Interface",
+        });
+
+        const refl = query(project, "HasDescriptor");
+        equal(refl.documents?.length, 1);
+
+        equal(refl.documents[0].content, [
+            { kind: "text", text: "External doc!" },
+        ]);
     });
 });