+++ src/test/renderer/testProject/src/mod2.ts
@@ -1,6 +1,0 @@
-export * from "./mod.js";
-
-/**
- * Will be exported from mod2, unlike the default function in mod
- */
-export default function () {}