+++ src/test/renderer/testProject/src/weird-names.ts
@@ -1,7 +1,0 @@
-/**
- * Object with characters that must be escaped.
- */
-export const foo = {
-    "<c-a>": "<c-b>",
-    "=": "=",
-};